modified Jenkinsfile and created tests directory
